refactor: refactor some components
